Final project code uploaded
This includes the finalised code and an Example folder which contains all test results from running the end-to-end protocol on the data as uploaded. The code was verified to run without errors prior to upload (A warning of passing a zero into a log function was thrown which is thought to be due to the list of testing energy levels/ MRR set. This did not appear to have an impact on the outcomes.